+++ avaframe/com2AB/com2AB.py
@@ -11,9 +11,9 @@ import numpy as np
 import matplotlib.pyplot as plt
 
 # Local imports
-import avaframe.in2Trans.geoTrans as geoTrans
+import avaframe.in3Utils.geoTrans as geoTrans
 import avaframe.in2Trans.shpConversion as shpConv
-import avaframe.in3Utils.ascUtils as IOf
+import avaframe.in2Trans.ascUtils as IOf
 
 # create local logger
 log = logging.getLogger(__name__)
