+++ avaframe/runCom2AB.py
@@ -4,8 +4,6 @@
 # Local imports
 from avaframe.com2AB import com2AB
 from avaframe.out3Plot import outAB
-import avaframe.in2Trans.shpConversion as shpConv
-import avaframe.in3Utils.ascUtils as IOf
 from avaframe.in3Utils import cfgUtils
 from avaframe.in3Utils import logUtils
