+++ avaframe/tests/test_shpConversion.py
@@ -4,7 +4,7 @@ import os
 import pytest
 
 # Local imports
-import avaframe.in3Utils.ascUtils as ascUtils
+import avaframe.in2Trans.ascUtils as ascUtils
 import avaframe.in2Trans.shpConversion as shpConv
